+++ samples/elaborate/contact_manager.rb
@@ -83,7 +83,7 @@ class ContactManager
           }
           items bind(@contact_manager_presenter, :results),
           column_properties(:first_name, :last_name, :email)
-          on_mouse_down { |event|
+          on_mouse_up { |event|
             table_proxy.edit_table_item(event.table_item, event.column_index)
           }
         }