+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support Glimmer::SWT::DNDProxy for handling Drop & Drop styles
 - Implemented list:* rake tasks for listing Glimmer custom widget gems, custom shell gems, and dsl gems
 - Fix bug with table edit remaining when sorting table or re-listing (in contact_manager.rb sample)
+- Update icon of scaffolded apps to Glimmer logo
 
 ## 0.1.2